Refactor common code in IntegerExperiment and MeasurementExperiment into StatisticsExperiment class
